ugc searcher fix

+++ searcher.py
@@ -8,11 +8,15 @@ from discord_webhook import DiscordWebhook, DiscordEmbed
 import subprocess
 import aiohttp
 import asyncio
+import json
 
-webhook_url = ""
+openjson = open('config.json') 
+conf = json.load(openjson)
 
-with open("cookie.txt", "r") as f:
-    cookie = f.read()
+webhook_url = conf["sniperWebhookURL"]
+cookie = conf["cookie"]
+
+snipedIds = []
 
 def betterPrint(text):
     now = time.strftime('%r')
@@ -32,6 +36,7 @@ def compare(list1, list2):
         return list2
     else:
         betterPrint(f"[violet]found {len(added_items)} new assets")
+        betterPrint(f"[violet]Assets: {added_items}")
         return added_items
 
 
@@ -45,10 +50,12 @@ def request_details(items, cookie):
         'cookie': f'.ROBLOSECURITY={cookie};',
         'x-csrf-token': get_x_token(cookie)
     }
-    payload = {'items': items}
+    #payload = {'items': items}
+    payload = {"items": [{"itemType": "Asset", "id": int(items)}]}
     response = requests.post('https://catalog.roblox.com/v1/catalog/items/details',
                              json=payload,
                              headers=headers)
+    
     return response.json()
 
 def extract_data(details):
@@ -73,20 +80,24 @@ def sendWebhook(val):
 
 
 async def main():
-    items = await latest()
-    betterPrint(f'loaded {len(items)} items')
+    iitems = await latest()
+    betterPrint(f'loaded {len(iitems)} items')
 
     while 1:
         new_items = await latest()
-        items = compare(new_items, items)
-
-        if len(items) < 100:
+        items = compare(new_items, iitems)
+        if items is not None and len(items) < 100 and items[0] not in iitems and items[0] not in snipedIds: 
             betterPrint('getting item info')
-            item_info = get_item_info(items, cookie)
-
+            item_info = get_item_info(items[0], cookie)
             for item in item_info:
-                sendWebhook(item)
-                if item.get('price', 'Offsale') == 0: #adds if lim is free
+                if conf["webhookEnabled"] == True:
+                    sendWebhook(item)
+                    print('sent webhook')
+                else:
+                    print(item)
+
+                snipedIds.append(item[id])
+                if int(item.get('price', 'Offsale')) == 0: #adds if lim is free
                     with open("limiteds.txt", "w") as file:
                         file.truncate()
                         file.write(f"{item['id']}")
@@ -95,18 +106,43 @@ async def main():
                         process.communicate(input=input_data.encode())
         await asyncio.sleep(5)
 
+
+headerss = {
+'Content-Type': 'application/json',
+'Accept': 'application/json',
+        }
+
+json_data = {
+'limit': 10,
+'cursor': '',
+'query': '',
+'catalogMode': 'All',
+'excludeNonCollectibles': True,
+'excludeReleased': False,
+'includeNotForSale': False,
+'includeForSale': True,
+'includeLimitedItems': True,
+'sortColumn': 'Updated',
+'sortOrder': 'Desc',
+'assetTypes': [],
+'assetGenres': [],
+'creators': [],
+'creatorsExclude': [],
+'minimumPrice': 0,
+'maximumPrice': 999999,
+'minimumOwnerCount': 0,
+'maximumOwnerCount': 0,
+             }
+
+
 async def latest():
-  try:
-    url = 'https://catalog.roblox.com/v2/search/items/details?itemRestrictions=Limited&Category=1&salesTypeFilter=2&SortType=3&Limit=120'
-    async with aiohttp.ClientSession() as session:
-        async with session.get(url) as response:
-            r = await response.json()
-            rData = r['data']
-            ids = []
-        for item in rData:
-            ids.append(item["id"])
+  async with aiohttp.ClientSession() as session:
+    async with session.post('https://rblx.trade/api/v2/catalog/asset-search/list', headers=headerss, json=json_data) as response:
+      r = await response.json(content_type='application/json')
+      rData = r['data']
+      ids = []
+    for item in rData:
+      ids.append(item["id"])
     return ids
-  except Exception as e:
-    print(e)
 
 asyncio.run(main())
